fix idea for "measure": SpectrometerServer: exception: Andor error in GetImages16: code 20069 bug
std::vector obj(res.dataSize / sizeof(int)); divide by sizeof(int) instead of sizeof(WORD) to get the correct number of elements in the vector.
try2fix no2

+++ .src/SpectrometerServer.cpp
@@ -137,7 +137,7 @@ void ProcessClient(HANDLE hPipe) {
                 int pixelsPerSpectrum = req.iArgs[1];
                 AppLogger::instance().info("SpectrometerServer: getAllSpectra");
                 auto data = cam.getAllSpectra(numSpectra, pixelsPerSpectrum);
-                res.dataSize = (int32_t)(data.size() * sizeof(WORD));
+                res.dataSize = (int32_t)(data.size() * sizeof(int));
                 
                 DWORD bytesWritten = 0;
                 if (!WriteFile(hPipe, &res, sizeof(res), &bytesWritten, NULL)) {
